Refactor OAuth App tests into a single spec
